+++ app/src/main/java/com/example/movieapp/data/dto/MovieDetailDto.kt
@@ -5,28 +5,20 @@ import com.fasterxml.jackson.annotation.JsonProperty
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class MovieDetailDto(
-    @JsonProperty("belongs_to_collection")
-    val belongsToCollection: CollectionDto?,
+    @JsonProperty("belongs_to_collection") val belongsToCollection: CollectionDto?,
     val genres: List<GenreDto>,
     val homepage: String?,
-    val id: Long,
-    @JsonProperty("imdb_id")
-    val imdbId: String?,
-    @JsonProperty("original_title")
-    val originalTitle: String,
+    val id: Int,
+    @JsonProperty("imdb_id") val imdbId: String?,
+    @JsonProperty("original_title") val originalTitle: String,
     val overview: String?,
     val popularity: Double,
-    @JsonProperty("poster_path")
-    val posterPath: String?,
-    @JsonProperty("production_companies")
-    val productionCompanies: List<ProductionCompanyDto>,
-    @JsonProperty("release_date")
-    val releaseDate: String?,
+    @JsonProperty("poster_path") val posterPath: String?,
+    @JsonProperty("production_companies") val productionCompanies: List<ProductionCompanyDto>,
+    @JsonProperty("release_date") val releaseDate: String?,
     val runtime: Int?,
     val tagline: String?,
     val title: String,
-    @JsonProperty("vote_average")
-    val voteAverage: Double,
-    @JsonProperty("vote_count")
-    val voteCount: Int
+    @JsonProperty("vote_average") val voteAverage: Double,
+    @JsonProperty("vote_count") val voteCount: Int
 )
